IOS-2542 Fix review
